etcdserver: Move version monitor logic to separate module
